ingestao-connectly/requirements.txt
google-cloud-storage
pandas
pyarrow
python-dotenv
google-cloud-bigquery
